FullStack_Docker-main/requirements.txt
# Required packages for the Book Inventory application
streamlit==1.10.0  # Specified version for consistency
psycopg2-binary==2.9.1  # Specified version for consistency
